requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python
albumentations
pyyaml
timm
einops
numpy
tqdm
tensorboard